fixed boot

+++ bootloader/src/bootloader.c
@@ -88,7 +88,7 @@ void handle_boot(void)
     uint32_t password_pos;
     uint32_t i = 0;
     uint8_t *rel_msg;
-    uint8_t *boot = (uint8_t *)FIRMWARE_BOOT_PTR;
+    uint8_t *boot = (uint32_t *)FIRMWARE_BOOT_PTR;
 
     // Acknowledge the host
     uart_writeb(HOST_UART, 'B');
